+++ source/custom_mp3player.h
@@ -16,7 +16,7 @@ void MP3Player_Volume(u32 volume);
 s32 MP3Player_PlayBuffer(const void *buffer,s32 len,bool (*filterfunc)(struct mad_stream *,struct mad_frame *));
 s32 MP3Player_PlayFile(void *cb_data,s32 (*reader)(void *,void *,s32),bool (*filterfunc)(struct mad_stream *,struct mad_frame *));
 float MP3Player_GetAmplitude(void);
-void MP3Player_SetSeconds(int seconds);
+void MP3Player_SetSeconds(float seconds);
 bool seek_filter(struct mad_stream *stream, struct mad_frame *frame);
 
 #ifdef __cplusplus
